+++ src/Methods/Sodium.php
@@ -178,12 +178,19 @@ class Sodium extends Method_Base {
 	/**
 	 * Encrypt a given string.
 	 *
+	 * @access private
+	 *
 	 * @param string $plain_text The plain string.
 	 *
 	 * @return string
 	 * @throws RuntimeException If error occurred.
 	 */
 	public function encrypt( string $plain_text ): string {
+		// bail if slug is not set.
+		if ( empty( $this->get_slug() ) ) {
+			return '';
+		}
+
 		// bail if it is unusable.
 		if ( ! $this->is_usable() ) {
 			return '';
@@ -220,6 +227,11 @@ class Sodium extends Method_Base {
 	 * @throws RuntimeException If error occurred.
 	 */
 	public function decrypt( string $encrypted_text ): string {
+		// bail if slug is not set.
+		if ( empty( $this->get_slug() ) ) {
+			return '';
+		}
+
 		// bail if it is unusable.
 		if ( ! $this->is_usable() ) {
 			return '';